requirements.txt
python-dotenv
fastapi==0.115.5
uvicorn[standard]
firebase-admin==6.5.0
langchain-community
langchain-google-genai
langchain-pinecone
langchain-core
pinecone-client
pypdf
langgraph
pydantic==2.10.3
google-genai==0.1.0
